INT-241 move more REs builders into separate files
